requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.28
botocore==1.34.28
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
packaging==23.2
phonenumbers==8.13.27
pillow==10.2.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pyscopg2==66.0.2
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
